Move classes like Irs1040Pdf into a module so they're not all just sitting in app/lib
Now they are PdfFiller::Irs1040Pdf etc
