+++ spec/lib/irs8812_ty2021_pdf_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Irs8812Ty2021Pdf do
+RSpec.describe PdfFiller::Irs8812Ty2021Pdf do
   include PdfSpecHelper
 
   let(:pdf) { described_class.new(submission) }
