+++ spec/lib/pdf_filler/irs1040_schedule_lep_pdf_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-describe Irs1040ScheduleLepPdf do
+describe PdfFiller::Irs1040ScheduleLepPdf do
   include PdfSpecHelper
   let(:pdf) { described_class.new(submission) }
   let(:submission) { create :efile_submission }
